+++ src/sphinx_argparse_cli/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 from .version import __version__
 
@@ -10,13 +10,15 @@ if TYPE_CHECKING:
     from sphinx.application import Sphinx
 
 
-def setup(app: Sphinx) -> None:
+def setup(app: Sphinx) -> dict[str, Any]:
     app.add_css_file("custom.css")
 
     from ._logic import SphinxArgparseCli
 
     app.add_directive(SphinxArgparseCli.name, SphinxArgparseCli)
     app.add_config_value("sphinx_argparse_cli_prefix_document", False, "env")  # noqa: FBT003
+
+    return {"parallel_read_safe": True}
 
 
 __all__ = [